requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
appnope==0.1.4
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.6.2
charset-normalizer==3.3.2
comm==0.2.2
dateparser==1.2.0
debugpy==1.8.2
decorator==5.1.1
exceptiongroup==1.2.1
executing==2.0.1
frozenlist==1.4.1
gevent==24.2.1
greenlet==3.0.3
idna==3.7
importlib-metadata==8.0.0
ipykernel==6.29.4
ipython==8.18.1
ipython-genutils==0.2.0
ipython-sql==0.5.0
jedi==0.19.1
jupyter-client==8.6.2
jupyter-core==5.7.2
matplotlib-inline==0.1.7
multidict==6.0.5
nest-asyncio==1.6.0
numpy==2.0.0
packaging==24.1
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
platformdirs==4.2.2
prettytable==3.10.0
prompt-toolkit==3.0.47
psutil==6.0.0
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pycryptodome==3.20.0
pygments==2.18.0
python-binance==1.0.19
python-dateutil==2.9.0.post0
pytz==2024.1
pyzmq==26.0.3
regex==2024.5.15
requests==2.32.3
six==1.16.0
SQLAlchemy==2.0.31
sqlparse==0.5.0
stack-data==0.6.3
tornado==6.4.1
traitlets==5.14.3
typing-extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
ujson==5.10.0
urllib3==2.2.2
wcwidth==0.2.13
websocket==0.2.1
websocket-client==1.8.0
websockets==12.0
yarl==1.9.4
zipp==3.19.2
zope.event==5.0
zope.interface==6.4.post2
